==> tasks/quora/01-LogisticRegression.ipynb <==
import numpy as np
import pandas as pd
import os

from sklearn.feature_extraction.text import HashingVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import precision_recall_fscore_support, accuracy_score

df = pd.read_csv('/Users/anton/mywork/Datasets/Quora/train.csv')

df.shape

df

df.shape[0], df[df.target == 1].shape[0], df[df.target == 1].shape[0] / df.shape[0]

X_train, X_test, y_train, y_test = train_test_split(df.question_text, df.target, test_size=0.2, random_state=42)

sum(y_train) / len(y_train)

sum(y_test) / len(y_test)

TODO: split more evenly

vectorizer = HashingVectorizer(n_features=2**15, norm=None, alternate_sign=False)
X_train_transformed = vectorizer.fit_transform(X_train)

print(X_train_transformed[1])

X_train[1]

X_test_transformed = vectorizer.transform(X_test)

log_clf = LogisticRegression(random_state=42)
log_clf.fit(X_train_transformed, y_train)

y_pred = log_clf.predict(X_test_transformed)

precision_recall_fscore_support(y_test, y_pred)

from sklearn.metrics import precision_score, recall_score
precision_score(y_test, y_pred), recall_score(y_test, y_pred)

accuracy_score(y_test, y_pred)
