==> tasks/quora/02-SplitData.ipynb <==
import numpy as np
import pandas as pd
import os

from sklearn.model_selection import StratifiedShuffleSplit

df = pd.read_csv('/Users/anton/mywork/Datasets/Quora/train.csv')

df.head()

df.shape

sum(df.target) / df.shape[0]

split = StratifiedShuffleSplit(n_splits=1, test_size=0.1, random_state=42)
for train_valid_index, test_index in split.split(df, df.target):
    train_valid_set = df.iloc[train_valid_index]
    test_set = df.iloc[test_index]

for train_index, valid_index in split.split(train_valid_set, train_valid_set.target):
    train_set = train_valid_set.iloc[train_index]
    valid_set = train_valid_set.iloc[valid_index]

split = StratifiedShuffleSplit(n_splits=1, test_size=0.2, random_state=42)
for train_index, test_valid_index in split.split(df, df.target):
    train_set = df.iloc[train_index]
    test_valid_set = df.iloc[test_valid_index]

split2 = StratifiedShuffleSplit(n_splits=1, test_size=0.5, random_state=42)
for test_index, valid_index in split2.split(test_valid_set, test_valid_set.target):
    test_set = test_valid_set.iloc[test_index]
    valid_set = test_valid_set.iloc[valid_index]

print(train_set.shape[0], sum(train_set.target) / train_set.shape[0])
print(valid_set.shape[0], sum(valid_set.target) / valid_set.shape[0])
print(test_set.shape[0], sum(test_set.target) / test_set.shape[0])

train_set.to_csv('/Users/anton/mywork/Datasets/Quora/split/train.csv', index=False)
valid_set.to_csv('/Users/anton/mywork/Datasets/Quora/split/valid.csv', index=False)
test_set.to_csv('/Users/anton/mywork/Datasets/Quora/split/test.csv', index=False)
