==> tasks/quora/03-LogisticRegression-Analysis.ipynb <==
import numpy as np
import pandas as pd
import os
import re

from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.feature_extraction.text import HashingVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import precision_recall_fscore_support

df = pd.read_csv('/Users/anton/mywork/Datasets/Quora/train.csv')

df.head()

split = StratifiedShuffleSplit(n_splits=1, test_size=0.2, random_state=42)
for train_index, test_valid_index in split.split(df, df.target):
    train_set = df.iloc[train_index]
    test_valid_set = df.iloc[test_valid_index]

split2 = StratifiedShuffleSplit(n_splits=1, test_size=0.5, random_state=42)
for test_index, valid_index in split2.split(test_valid_set, test_valid_set.target):
    test_set = test_valid_set.iloc[test_index]
    valid_set = test_valid_set.iloc[valid_index]

print("Train set", train_set.shape[0], sum(train_set.target) / train_set.shape[0])
print("Valid set", valid_set.shape[0], sum(valid_set.target) / valid_set.shape[0])
print("Test set", test_set.shape[0], sum(test_set.target) / test_set.shape[0])

vectorizer = HashingVectorizer(n_features=2**15, norm=None, alternate_sign=False)
X_train_transformed = vectorizer.fit_transform(train_set.question_text)
y_train = train_set.target

X_valid_transformed = vectorizer.transform(valid_set.question_text)
y_valid = valid_set.target

log_clf = LogisticRegression(random_state=42)
log_clf.fit(X_train_transformed, y_train)

y_pred = log_clf.predict(X_valid_transformed)

precision_recall_fscore_support(y_valid, y_pred)

y_train_pred = log_clf.predict(X_train_transformed)
precision_recall_fscore_support(y_train, y_train_pred)

train_set.question_text.iloc[0]

print(X_train_transformed[0])

test_text = 'Which is the best career after 12th?'
print(vectorizer.transform([test_text]))

test_text = 'career'
t = vectorizer.transform([test_text])
t.indices[0]

log_clf.coef_[0].shape

test_text = 'Which is the best career after 12th?'
re.findall(r'\b\w\w+\b', test_text)

words = set([word.lower() for question in train_set.question_text for word in re.findall(r'\b\w\w+\b', question)])

len(words)

m = {}
for word in words:
    transformed = vectorizer.transform([word])
    if transformed.nnz == 1:
        code = transformed.indices[0]
        if code in m:
            m[code] += ',' + word
        else:
            m[code] = word


for idx in np.flip(np.argsort(log_clf.coef_[0]), axis=0)[:30]:
    print(log_clf.coef_[0][idx], m[idx])

for idx in np.argsort(log_clf.coef_[0])[:30]:
    print(log_clf.coef_[0][idx], m[idx])

X_train_transformed

X_train_transformed.nnz

Roughly 120MB
